Start splitting out a separate sys package
Moved array and base so far
