+++ .gitignore
@@ -1,3 +1,3 @@
-/Cargo.lock
-/target
+Cargo.lock
+target/
