+++ core-foundation-sys/Cargo.toml
@@ -1,12 +1,13 @@
 [package]
-name = "core-foundation"
+name = "core-foundation-sys"
 description = "Bindings to Core Foundation for OS X"
 homepage = "https://github.com/servo/core-foundation-rs"
 repository = "https://github.com/servo/core-foundation-rs"
 version = "0.1.0"
 authors = ["The Servo Project Developers"]
 license = "MIT / Apache-2.0"
+build = "build.rs"
+links = "CoreFoundation.framework"
 
 [dependencies]
 libc = "0.1"
-core-foundation-sys = { path = "../core-foundation-sys" }
