+++ core-foundation-sys/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    println!("cargo:rustc-link-lib=framework=CoreFoundation");
+}
