+++ core-foundation-sys/src/lib.rs
@@ -1,0 +1,6 @@
+#![allow(non_snake_case, non_camel_case_types)]
+
+extern crate libc;
+
+pub mod array;
+pub mod base;
